Refactoring backend
Put command in seperate file
Deleted empty models.py
Removed comment from postgres_tools
