+++ server/app.py
@@ -14,9 +14,18 @@ from flask_sqlalchemy import SQLAlchemy
 from flask_marshmallow import Marshmallow
 
 from postgres_tools import Connection as PgConnection
+from command import db as command_db, ma as command_ma, Command, command_bpr
 
 
 app = Flask(__name__, static_url_path='')
+
+# Registering command blueprint incl. sqlalchemy and marshmallow
+app.register_blueprint(command_bpr, url_prefix="/command")
+command_db.init_app(app)
+command_ma.init_app(app)
+
+
+
 CORS(app)
 socketio = SocketIO(app, cors_allowed_origins='*')
 
@@ -24,6 +33,7 @@ app.config['JSON_AS_ASCII'] = False
 basedir = os.path.abspath(os.path.dirname(__file__))
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///' + os.path.join(basedir, 'db.sqlite')
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+
 
 db = SQLAlchemy(app)
 ma = Marshmallow(app)
@@ -77,18 +87,6 @@ def notify(payload):
     event_index += 1
 
 
-class Command(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(255))
-    query_string = db.Column(db.String())
-    severity = db.Column(db.String(255))
-
-    def __init__(self, name, query_string, severity):
-        self.name = name
-        self.query_string = query_string
-        self.severity = severity
-
-
 def events(connection):
     con = connection.con
     while True:
@@ -112,12 +110,6 @@ class ConnectionSchema(ma.Schema):
         fields=('id', 'name', 'database', 'user', 'password', 'host', 'port')
 connection_schema = ConnectionSchema()
 connections_schema = ConnectionSchema(many=True)
-
-class CommandSchema(ma.Schema):
-    class Meta:
-        fields=('id', 'name', 'query_string', 'severity')
-command_schema = CommandSchema()
-commands_schema = CommandSchema(many=True)
 
 
 def is_connection_valid(connection):
@@ -286,24 +278,6 @@ def execute_command():
     response = connection.get_connection().execute_command(db_query)
     
     return jsonify(response)
-
-
-@app.route("/command", methods=["POST"])
-def create_query():
-    name = request.json["name"]
-    query_string = request.json["query_string"]
-    severity = request.json["severity"]
-
-    new_command = Command(name, query_string, severity)
-    db.session.add(new_command)
-    db.session.commit()
-    return command_schema.jsonify(new_command)
-
-@app.route('/command/all', methods=['GET'])
-def get_all_commands():
-    all_commands = Command.query.all()
-    return commands_schema.jsonify(all_commands)
-
 
 
 @app.route("/connection/<int:id>/trigger", methods=["POST"])
